feat: add constants for homepage Perpetuals section and refactor Perps components
