+++ app/components/Views/Homepage/Sections/Perpetuals/HomepagePerpsHomeSlot.test.tsx
@@ -24,69 +24,41 @@ jest.mock('../../../../../hooks', () => ({
     Reflect.apply(mockUseABTest, undefined, args),
 }));
 
-jest.mock('../../../../UI/Perps/hooks', () => ({
-  usePerpsLivePositions: jest.fn(),
-  usePerpsLiveOrders: jest.fn(),
-}));
-
-jest.mock('../../../../UI/Perps/hooks/usePerpsConnection', () => ({
-  usePerpsConnection: jest.fn(),
-}));
-
 jest.mock('./PerpsSection', () => {
   const ReactLib = jest.requireActual('react');
   const RN = jest.requireActual('react-native');
   return {
     __esModule: true,
-    default: ReactLib.forwardRef((_props: unknown, _ref: unknown) =>
-      ReactLib.createElement(RN.Text, null, 'PerpsSection'),
+    default: ReactLib.forwardRef(
+      (
+        props: {
+          emptyStateContent?: string;
+          emptyStateTitleOverride?: string;
+        },
+        _ref: unknown,
+      ) =>
+        ReactLib.createElement(
+          RN.View,
+          null,
+          ReactLib.createElement(RN.Text, null, 'PerpsSection'),
+          ReactLib.createElement(
+            RN.Text,
+            null,
+            `emptyStateContent:${props.emptyStateContent ?? 'tiles'}`,
+          ),
+          ReactLib.createElement(
+            RN.Text,
+            null,
+            `emptyStateTitle:${props.emptyStateTitleOverride ?? 'default'}`,
+          ),
+        ),
     ),
   };
 });
-
-jest.mock('./HomepagePerpsMoversSection', () => {
-  const ReactLib = jest.requireActual('react');
-  const RN = jest.requireActual('react-native');
-  return {
-    __esModule: true,
-    default: ReactLib.forwardRef((_props: unknown, _ref: unknown) =>
-      ReactLib.createElement(RN.Text, null, 'HomepagePerpsMoversSection'),
-    ),
-  };
-});
-
-const mockUsePerpsLivePositions = jest.mocked(
-  jest.requireMock('../../../../UI/Perps/hooks').usePerpsLivePositions,
-);
-const mockUsePerpsLiveOrders = jest.mocked(
-  jest.requireMock('../../../../UI/Perps/hooks').usePerpsLiveOrders,
-);
-const mockUsePerpsConnection = jest.mocked(
-  jest.requireMock('../../../../UI/Perps/hooks/usePerpsConnection')
-    .usePerpsConnection,
-);
 
 describe('HomepagePerpsHomeSlot', () => {
   beforeEach(() => {
     jest.clearAllMocks();
-    mockUsePerpsLivePositions.mockReturnValue({
-      positions: [],
-      isInitialLoading: false,
-    });
-    mockUsePerpsLiveOrders.mockReturnValue({
-      orders: [],
-      isInitialLoading: false,
-    });
-    mockUsePerpsConnection.mockReturnValue({
-      isConnected: true,
-      isConnecting: false,
-      isInitialized: true,
-      error: null,
-      connect: jest.fn(),
-      disconnect: jest.fn(),
-      resetError: jest.fn(),
-      reconnectWithNewContext: jest.fn(),
-    });
   });
 
   it('renders PerpsSection when experiment is control', () => {
@@ -113,12 +85,11 @@ describe('HomepagePerpsHomeSlot', () => {
     );
 
     expect(screen.getByText('PerpsSection')).toBeOnTheScreen();
-    expect(screen.queryByText('HomepagePerpsMoversSection')).toBeNull();
-    expect(mockUsePerpsLivePositions).not.toHaveBeenCalled();
-    expect(mockUsePerpsLiveOrders).not.toHaveBeenCalled();
+    expect(screen.getByText('emptyStateContent:tiles')).toBeOnTheScreen();
+    expect(screen.getByText('emptyStateTitle:default')).toBeOnTheScreen();
   });
 
-  it('renders HomepagePerpsMoversSection when experiment is treatment and user has no positions or orders', () => {
+  it('asks PerpsSection to render pills for the empty state when experiment is treatment', () => {
     mockUseABTest.mockImplementation((key: string) => {
       if (key === HOMEPAGE_PERPS_PILLS_EMPTY_AB_KEY) {
         return {
@@ -135,45 +106,6 @@ describe('HomepagePerpsHomeSlot', () => {
         variantName: 'control',
         isActive: false,
       };
-    });
-
-    renderWithProvider(
-      <HomepagePerpsHomeSlot sectionIndex={1} totalSectionsLoaded={5} />,
-    );
-
-    expect(screen.getByText('HomepagePerpsMoversSection')).toBeOnTheScreen();
-    expect(screen.queryByText('PerpsSection')).toBeNull();
-    expect(mockUsePerpsLivePositions).toHaveBeenCalled();
-    expect(mockUsePerpsLiveOrders).toHaveBeenCalled();
-  });
-
-  it('renders PerpsSection when treatment user is empty and connection is errored', () => {
-    mockUseABTest.mockImplementation((key: string) => {
-      if (key === HOMEPAGE_PERPS_PILLS_EMPTY_AB_KEY) {
-        return {
-          variant:
-            HOMEPAGE_PERPS_PILLS_EMPTY_VARIANTS[
-              HomepagePerpsPillsEmptyVariant.Treatment
-            ],
-          variantName: HomepagePerpsPillsEmptyVariant.Treatment,
-          isActive: true,
-        };
-      }
-      return {
-        variant: {},
-        variantName: 'control',
-        isActive: false,
-      };
-    });
-    mockUsePerpsConnection.mockReturnValue({
-      isConnected: false,
-      isConnecting: false,
-      isInitialized: false,
-      error: 'CONNECTION_TIMEOUT',
-      connect: jest.fn(),
-      disconnect: jest.fn(),
-      resetError: jest.fn(),
-      reconnectWithNewContext: jest.fn(),
     });
 
     renderWithProvider(
@@ -181,7 +113,8 @@ describe('HomepagePerpsHomeSlot', () => {
     );
 
     expect(screen.getByText('PerpsSection')).toBeOnTheScreen();
-    expect(screen.queryByText('HomepagePerpsMoversSection')).toBeNull();
+    expect(screen.getByText('emptyStateContent:pills')).toBeOnTheScreen();
+    expect(screen.getByText('emptyStateTitle:Perps movers')).toBeOnTheScreen();
   });
 
   it('renders PerpsSection when mode is positions-only even in treatment', () => {
@@ -212,38 +145,6 @@ describe('HomepagePerpsHomeSlot', () => {
     );
 
     expect(screen.getByText('PerpsSection')).toBeOnTheScreen();
-    expect(mockUsePerpsLivePositions).not.toHaveBeenCalled();
-    expect(mockUsePerpsLiveOrders).not.toHaveBeenCalled();
-  });
-
-  it('renders PerpsSection when user has open positions', () => {
-    mockUseABTest.mockImplementation((key: string) => {
-      if (key === HOMEPAGE_PERPS_PILLS_EMPTY_AB_KEY) {
-        return {
-          variant:
-            HOMEPAGE_PERPS_PILLS_EMPTY_VARIANTS[
-              HomepagePerpsPillsEmptyVariant.Treatment
-            ],
-          variantName: HomepagePerpsPillsEmptyVariant.Treatment,
-          isActive: true,
-        };
-      }
-      return {
-        variant: {},
-        variantName: 'control',
-        isActive: false,
-      };
-    });
-    mockUsePerpsLivePositions.mockReturnValue({
-      positions: [{ symbol: 'ETH-PERP' } as never],
-      isInitialLoading: false,
-    });
-
-    renderWithProvider(
-      <HomepagePerpsHomeSlot sectionIndex={1} totalSectionsLoaded={5} />,
-    );
-
-    expect(screen.getByText('PerpsSection')).toBeOnTheScreen();
-    expect(mockUsePerpsLivePositions).toHaveBeenCalled();
+    expect(screen.getByText('emptyStateContent:tiles')).toBeOnTheScreen();
   });
 });
