+++ app/components/Views/Homepage/Sections/Perpetuals/HomepagePerpsHomeSlot.tsx
@@ -5,62 +5,15 @@ import {
   HOMEPAGE_PERPS_PILLS_EMPTY_AB_TEST_EXPOSURE_OPTIONS,
   HOMEPAGE_PERPS_PILLS_EMPTY_VARIANTS,
 } from '../../abTestConfig';
-import {
-  usePerpsLiveOrders,
-  usePerpsLivePositions,
-} from '../../../../UI/Perps/hooks';
-import { usePerpsConnection } from '../../../../UI/Perps/hooks/usePerpsConnection';
 import type { SectionRefreshHandle } from '../../types';
 import type { PerpsSectionProps } from './PerpsSectionWithProvider';
 import PerpsSection from './PerpsSection';
-import HomepagePerpsMoversSection from './HomepagePerpsMoversSection';
-
-const HOMEPAGE_THROTTLE_MS = 5000;
+import { strings } from '../../../../../../locales/i18n';
 
 /**
- * Only mounted when the TMCU-725 experiment is in treatment (and not
- * positions-only). Owns `usePerpsLivePositions` / `usePerpsLiveOrders` so control
- * users never pay for duplicate subscriptions with `PerpsSection`.
- */
-const HomepagePerpsTreatmentEmptyBranch = forwardRef<
-  SectionRefreshHandle,
-  PerpsSectionProps
->((props, ref) => {
-  const { positions, isInitialLoading: positionsLoading } =
-    usePerpsLivePositions({
-      throttleMs: HOMEPAGE_THROTTLE_MS,
-    });
-
-  const { orders, isInitialLoading: ordersLoading } = usePerpsLiveOrders({
-    hideTpSl: true,
-    throttleMs: HOMEPAGE_THROTTLE_MS,
-  });
-
-  const { error: connectionError } = usePerpsConnection();
-
-  const showSkeleton = positionsLoading || ordersLoading;
-  const hasItems = positions.length > 0 || orders.length > 0;
-
-  const showMoversEmpty = !showSkeleton && !hasItems;
-
-  if (connectionError) {
-    return <PerpsSection ref={ref} {...props} />;
-  }
-
-  if (showMoversEmpty) {
-    return <HomepagePerpsMoversSection ref={ref} {...props} />;
-  }
-
-  return <PerpsSection ref={ref} {...props} />;
-});
-
-HomepagePerpsTreatmentEmptyBranch.displayName =
-  'HomepagePerpsTreatmentEmptyBranch';
-
-/**
- * Chooses between the classic homepage Perps section and the Explore-style
- * "Perps movers" rail for the empty state when the TMCU-725 experiment is in treatment.
- * `PerpsSection.tsx` stays unchanged; experiment wiring lives here.
+ * Chooses the empty-state content for the homepage Perps section.
+ * The section shell, header, navigation, analytics, and refresh contract all
+ * stay owned by `PerpsSection`.
  */
 const HomepagePerpsHomeSlot = forwardRef<
   SectionRefreshHandle,
@@ -82,7 +35,14 @@ const HomepagePerpsHomeSlot = forwardRef<
     return <PerpsSection ref={ref} {...props} />;
   }
 
-  return <HomepagePerpsTreatmentEmptyBranch ref={ref} {...props} />;
+  return (
+    <PerpsSection
+      ref={ref}
+      {...props}
+      emptyStateContent="pills"
+      emptyStateTitleOverride={strings('trending.perps_movers')}
+    />
+  );
 });
 
 HomepagePerpsHomeSlot.displayName = 'HomepagePerpsHomeSlot';
