+++ app/components/Views/Homepage/Sections/Perpetuals/PerpsSection.test.tsx
@@ -30,11 +30,22 @@ const mockUseHomepagePerpsPillsEmptyTransactionActiveAbTests = jest.fn<
   []
 >(() => undefined);
 
+const mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook = jest.fn(
+  (isHomepagePerpsPillsEmptySurface: boolean) =>
+    isHomepagePerpsPillsEmptySurface
+      ? mockUseHomepagePerpsPillsEmptyTransactionActiveAbTests()
+      : undefined,
+);
+
 jest.mock(
   '../../hooks/useHomepagePerpsPillsEmptyTransactionActiveAbTests',
   () => ({
-    useHomepagePerpsPillsEmptyTransactionActiveAbTests: () =>
-      mockUseHomepagePerpsPillsEmptyTransactionActiveAbTests(),
+    useHomepagePerpsPillsEmptyTransactionActiveAbTests: (
+      isHomepagePerpsPillsEmptySurface: boolean,
+    ) =>
+      mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook(
+        isHomepagePerpsPillsEmptySurface,
+      ),
   }),
 );
 
@@ -202,6 +213,27 @@ jest.mock('./components/PerpsMarketTileCard', () => {
   };
 });
 
+jest.mock('../../../TrendingView/feeds/perps/PerpsPillItem', () => {
+  const { TouchableOpacity, Text } = jest.requireActual('react-native');
+  return {
+    __esModule: true,
+    default: ({
+      item,
+      onNavigateToMarketDetails,
+    }: {
+      item: { market: { symbol: string } };
+      onNavigateToMarketDetails?: (market: { symbol: string }) => void;
+    }) => (
+      <TouchableOpacity
+        testID={`homepage-perps-pill-${item.market.symbol}`}
+        onPress={() => onNavigateToMarketDetails?.(item.market)}
+      >
+        <Text>{`pill-${item.market.symbol}`}</Text>
+      </TouchableOpacity>
+    ),
+  };
+});
+
 const {
   usePerpsLivePositions,
   usePerpsLiveOrders,
@@ -278,6 +310,14 @@ jest.mock('../../hooks/useHomeViewedEvent', () => ({
   },
 }));
 
+const spyOnUsePerpsFeed = () =>
+  jest.spyOn(
+    jest.requireActual(
+      '../../../TrendingView/feeds/perps/usePerpsFeed',
+    ) as typeof import('../../../TrendingView/feeds/perps/usePerpsFeed'),
+    'usePerpsFeed',
+  );
+
 describe('PerpsSection', () => {
   beforeEach(() => {
     jest.clearAllMocks();
@@ -316,6 +356,7 @@ describe('PerpsSection', () => {
     mockUseHomepagePerpsPillsEmptyTransactionActiveAbTests.mockReturnValue(
       undefined,
     );
+    mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook.mockClear();
   });
 
   it('renders section title', () => {
@@ -417,7 +458,7 @@ describe('PerpsSection', () => {
     });
   });
 
-  it('carries empty-surface active_ab_tests to perps home on title press', () => {
+  it('does not attach pills empty-state transaction AB tests on title press for tiles empty state', () => {
     const activeAbTests = [
       createActiveABTestAssignment(
         'homeTMCU725AbtestHomepagePerpsPillsEmptyState',
@@ -432,14 +473,15 @@ describe('PerpsSection', () => {
       <PerpsSection sectionIndex={0} totalSectionsLoaded={1} />,
     );
 
+    expect(
+      mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook,
+    ).toHaveBeenCalledWith(false);
+
     fireEvent.press(screen.getByText('Perpetuals'));
 
     expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
       screen: Routes.PERPS.PERPS_HOME,
-      params: {
-        source: 'home_section',
-        transactionActiveAbTests: activeAbTests,
-      },
+      params: { source: 'home_section' },
     });
   });
 
@@ -680,18 +722,6 @@ describe('PerpsSection', () => {
   });
 
   describe('Trending Perps Carousel', () => {
-    const makeTrendingMarket = (overrides: Record<string, unknown> = {}) => ({
-      symbol: 'BTC',
-      name: 'Bitcoin',
-      maxLeverage: '50x',
-      price: '$52,000',
-      change24h: '+$2,000',
-      change24hPercent: '+4.00%',
-      volume: '$2.5B',
-      volumeNumber: 2500000000,
-      ...overrides,
-    });
-
     it('shows trending carousel when user has no positions or orders', () => {
       usePerpsMarkets.mockReturnValue({
         markets: [
@@ -763,7 +793,7 @@ describe('PerpsSection', () => {
       expect(screen.getByTestId('homepage-perps-positions')).toBeOnTheScreen();
     });
 
-    it('navigates to market details and tracks active_ab_tests when tile is pressed', () => {
+    it('does not attach pills empty-state transaction AB tests when a carousel tile is pressed', () => {
       const market = makeTrendingMarket({ symbol: 'SOL' });
       const activeAbTests = [
         createActiveABTestAssignment(
@@ -786,6 +816,10 @@ describe('PerpsSection', () => {
         <PerpsSection sectionIndex={0} totalSectionsLoaded={1} />,
       );
 
+      expect(
+        mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook,
+      ).toHaveBeenCalledWith(false);
+
       fireEvent.press(screen.getByTestId('perps-market-tile-SOL'));
 
       expect(mockTrack).toHaveBeenCalledWith(
@@ -797,8 +831,10 @@ describe('PerpsSection', () => {
             PERPS_EVENT_VALUE.BUTTON_CLICKED.OPEN_POSITION,
           [PERPS_EVENT_PROPERTY.BUTTON_LOCATION]:
             PERPS_EVENT_VALUE.BUTTON_LOCATION.WALLET_HOME,
-          active_ab_tests: activeAbTests,
         }),
+      );
+      expect(mockTrack.mock.calls.at(-1)?.[1]).not.toHaveProperty(
+        'active_ab_tests',
       );
 
       expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
@@ -806,12 +842,11 @@ describe('PerpsSection', () => {
         params: {
           market,
           source: 'home_section',
-          transactionActiveAbTests: activeAbTests,
         },
       });
     });
 
-    it('merges trending and empty-surface active_ab_tests on carousel tile press', () => {
+    it('merges trending active_ab_tests on carousel tile press without pills empty-state attribution', () => {
       const market = makeTrendingMarket({ symbol: 'SOL' });
       const trendingAbTests = [
         createActiveABTestAssignment(
@@ -845,11 +880,10 @@ describe('PerpsSection', () => {
 
       fireEvent.press(screen.getByTestId('perps-market-tile-SOL'));
 
-      const merged = [...trendingAbTests, ...perpsEmptyAbTests];
       expect(mockTrack).toHaveBeenCalledWith(
         MetaMetricsEvents.PERPS_UI_INTERACTION,
         expect.objectContaining({
-          active_ab_tests: merged,
+          active_ab_tests: trendingAbTests,
         }),
       );
       expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
@@ -857,7 +891,7 @@ describe('PerpsSection', () => {
         params: {
           market,
           source: 'home_section',
-          transactionActiveAbTests: merged,
+          transactionActiveAbTests: trendingAbTests,
         },
       });
     });
@@ -932,7 +966,7 @@ describe('PerpsSection', () => {
       });
     });
 
-    it('carries empty-surface active_ab_tests when "View more" opens market list', () => {
+    it('does not attach pills empty-state transaction AB tests when "View more" opens market list', () => {
       const activeAbTests = [
         createActiveABTestAssignment(
           'homeTMCU725AbtestHomepagePerpsPillsEmptyState',
@@ -960,10 +994,7 @@ describe('PerpsSection', () => {
 
       expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
         screen: Routes.PERPS.MARKET_LIST,
-        params: {
-          source: 'home_section',
-          transactionActiveAbTests: activeAbTests,
-        },
+        params: { source: 'home_section' },
       });
     });
 
@@ -981,19 +1012,335 @@ describe('PerpsSection', () => {
     });
   });
 
-  describe('Watchlist Markets in Carousel', () => {
-    const makeTrendingMarket = (overrides: Record<string, unknown> = {}) => ({
-      symbol: 'BTC',
-      name: 'Bitcoin',
-      maxLeverage: '50x',
-      price: '$52,000',
-      change24h: '+$2,000',
-      change24hPercent: '+4.00%',
-      volume: '$2.5B',
-      volumeNumber: 2500000000,
-      ...overrides,
+  describe('pills empty state content', () => {
+    it('skips pills feed fetch when the user has positions', () => {
+      const usePerpsFeedSpy = spyOnUsePerpsFeed();
+      usePerpsLivePositions.mockReturnValue({
+        positions: [makePosition()],
+        isInitialLoading: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+        />,
+      );
+
+      expect(usePerpsFeedSpy).toHaveBeenCalledWith(
+        expect.objectContaining({ skipInitialFetch: true }),
+      );
+      usePerpsFeedSpy.mockRestore();
     });
 
+    it('fetches pills feed when the pills empty state is showing', () => {
+      const usePerpsFeedSpy = spyOnUsePerpsFeed();
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+        />,
+      );
+
+      expect(usePerpsFeedSpy).toHaveBeenCalledWith(
+        expect.objectContaining({ skipInitialFetch: false }),
+      );
+      usePerpsFeedSpy.mockRestore();
+    });
+
+    it('activates pills transaction AB hook only when pills empty state is showing', () => {
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+        />,
+      );
+
+      expect(
+        mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook,
+      ).toHaveBeenCalledWith(true);
+    });
+
+    it('carries pills empty-state transaction AB tests on title press', () => {
+      const activeAbTests = [
+        createActiveABTestAssignment(
+          'homeTMCU725AbtestHomepagePerpsPillsEmptyState',
+          'treatment',
+        ),
+      ];
+      mockUseHomepagePerpsPillsEmptyTransactionActiveAbTests.mockReturnValue(
+        activeAbTests,
+      );
+      usePerpsMarkets.mockReturnValue({
+        markets: [
+          makeTrendingMarket({ symbol: 'BTC', volumeNumber: 5000000000 }),
+        ],
+        isLoading: false,
+        error: null,
+        refresh: jest.fn(),
+        isRefreshing: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      fireEvent.press(screen.getByText('Perps movers'));
+
+      expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
+        screen: Routes.PERPS.PERPS_HOME,
+        params: {
+          source: 'home_section',
+          transactionActiveAbTests: activeAbTests,
+        },
+      });
+    });
+
+    it('returns null when pills feed is not loading and has no markets', () => {
+      usePerpsMarkets.mockReturnValue({
+        markets: [],
+        isLoading: false,
+        error: null,
+        refresh: jest.fn(),
+        isRefreshing: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      expect(screen.queryByText('Perps movers')).toBeNull();
+      expect(screen.toJSON()).toBeNull();
+    });
+
+    it('renders pills inside the shared section shell when configured', () => {
+      usePerpsMarkets.mockReturnValue({
+        markets: [
+          makeTrendingMarket({ symbol: 'BTC', volumeNumber: 5000000000 }),
+        ],
+        isLoading: false,
+        error: null,
+        refresh: jest.fn(),
+        isRefreshing: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      expect(screen.getByText('Perps movers')).toBeOnTheScreen();
+      expect(screen.getByTestId('homepage-perps-pills-list')).toBeOnTheScreen();
+      expect(screen.getByText('pill-BTC')).toBeOnTheScreen();
+      expect(
+        screen.queryByTestId('homepage-trending-perps-carousel'),
+      ).toBeNull();
+    });
+
+    it('keeps the shared header navigation for the pills section title', () => {
+      usePerpsMarkets.mockReturnValue({
+        markets: [
+          makeTrendingMarket({ symbol: 'BTC', volumeNumber: 5000000000 }),
+        ],
+        isLoading: false,
+        error: null,
+        refresh: jest.fn(),
+        isRefreshing: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      expect(
+        mockUseHomepagePerpsPillsEmptyTransactionActiveAbTestsHook,
+      ).toHaveBeenCalledWith(true);
+
+      fireEvent.press(screen.getByText('Perps movers'));
+
+      expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
+        screen: Routes.PERPS.PERPS_HOME,
+        params: { source: 'home_section' },
+      });
+    });
+
+    it('keeps the normal section title when positions are present', () => {
+      usePerpsLivePositions.mockReturnValue({
+        positions: [makePosition()],
+        isInitialLoading: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      expect(screen.getByText('Perpetuals')).toBeOnTheScreen();
+      expect(screen.queryByText('Perps movers')).toBeNull();
+      expect(screen.getByTestId('homepage-perps-positions')).toBeOnTheScreen();
+    });
+
+    it('uses the pills feed market data for position navigation when the tile feed is skipped', () => {
+      const fullMarket = {
+        symbol: 'BTC',
+        maxLeverage: 50,
+        marketType: 'crypto',
+        marketSource: 'HyperLiquid',
+        volumeNumber: 1000000,
+      };
+      usePerpsLivePositions.mockReturnValue({
+        positions: [makePosition()],
+        isInitialLoading: false,
+      });
+      usePerpsMarkets
+        .mockReturnValueOnce({
+          markets: [],
+          isLoading: false,
+          error: null,
+          refresh: jest.fn(),
+          isRefreshing: false,
+        })
+        .mockReturnValueOnce({
+          markets: [fullMarket],
+          isLoading: false,
+          error: null,
+          refresh: jest.fn(),
+          isRefreshing: false,
+        });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      fireEvent.press(screen.getByTestId('perps-position-row-BTC'));
+
+      expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
+        screen: Routes.PERPS.MARKET_DETAILS,
+        params: {
+          market: fullMarket,
+          initialTab: 'position',
+          source: 'section_position',
+        },
+      });
+    });
+
+    it('tracks and navigates through the shared market-details handler on pill press', () => {
+      const market = makeTrendingMarket({
+        symbol: 'SOL',
+        volumeNumber: 5000000000,
+      });
+      const activeAbTests = [
+        createActiveABTestAssignment(
+          'homeTMCU725AbtestHomepagePerpsPillsEmptyState',
+          'treatment',
+        ),
+      ];
+      mockUseHomepagePerpsPillsEmptyTransactionActiveAbTests.mockReturnValue(
+        activeAbTests,
+      );
+      usePerpsMarkets.mockReturnValue({
+        markets: [market],
+        isLoading: false,
+        error: null,
+        refresh: jest.fn(),
+        isRefreshing: false,
+      });
+
+      renderWithProvider(
+        <PerpsSection
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      fireEvent.press(screen.getByTestId('homepage-perps-pill-SOL'));
+
+      expect(mockTrack).toHaveBeenCalledWith(
+        MetaMetricsEvents.PERPS_UI_INTERACTION,
+        expect.objectContaining({
+          [PERPS_EVENT_PROPERTY.INTERACTION_TYPE]:
+            PERPS_EVENT_VALUE.INTERACTION_TYPE.BUTTON_CLICKED,
+          [PERPS_EVENT_PROPERTY.BUTTON_CLICKED]:
+            PERPS_EVENT_VALUE.BUTTON_CLICKED.OPEN_POSITION,
+          [PERPS_EVENT_PROPERTY.BUTTON_LOCATION]:
+            PERPS_EVENT_VALUE.BUTTON_LOCATION.WALLET_HOME,
+          active_ab_tests: activeAbTests,
+        }),
+      );
+      expect(mockNavigate).toHaveBeenCalledWith(Routes.PERPS.ROOT, {
+        screen: Routes.PERPS.MARKET_DETAILS,
+        params: {
+          market,
+          source: 'home_section',
+          transactionActiveAbTests: activeAbTests,
+        },
+      });
+    });
+
+    it('refreshes the pills feed while the pills empty state is showing', async () => {
+      const refresh = jest.fn().mockResolvedValue(undefined);
+      usePerpsMarkets.mockReturnValue({
+        markets: [
+          makeTrendingMarket({ symbol: 'BTC', volumeNumber: 5000000000 }),
+        ],
+        isLoading: false,
+        error: null,
+        refresh,
+        isRefreshing: false,
+      });
+      const ref = React.createRef<{ refresh: () => Promise<void> }>();
+
+      renderWithProvider(
+        <PerpsSection
+          ref={ref}
+          sectionIndex={0}
+          totalSectionsLoaded={1}
+          emptyStateContent="pills"
+          emptyStateTitleOverride="Perps movers"
+        />,
+      );
+
+      await ref.current?.refresh();
+
+      expect(refresh).toHaveBeenCalledTimes(1);
+    });
+  });
+
+  describe('Watchlist Markets in Carousel', () => {
     const watchlistState = (symbols: string[]) => ({
       engine: {
         backgroundState: {
