+++ app/components/Views/Homepage/Sections/Perpetuals/PerpsSectionTrendingOnly.tsx
@@ -1,0 +1,107 @@
+import React, { forwardRef, useImperativeHandle, useMemo, useRef } from 'react';
+import { View } from 'react-native';
+import { Box } from '@metamask/design-system-react-native';
+import SectionHeader from '../../../../../component-library/components-temp/SectionHeader';
+import SectionRow from '../../components/SectionRow';
+import PerpsPositionSkeleton from './components/PerpsPositionSkeleton';
+import PerpsTrendingCarousel from './components/PerpsTrendingCarousel';
+import { useHomepageSparklines } from './hooks/useHomepageSparklines';
+import { usePerpsTrendingCarouselData } from './hooks/usePerpsTrendingCarouselData';
+import { strings } from '../../../../../../locales/i18n';
+import type { SectionRefreshHandle } from '../../types';
+import useHomeViewedEvent, {
+  HomeSectionNames,
+} from '../../hooks/useHomeViewedEvent';
+import type { PerpsSectionProps } from './PerpsSectionWithProvider';
+import { useHomepageTrendingTransactionActiveAbTests } from '../../hooks/useHomepageTrendingTransactionActiveAbTests';
+import { WalletViewSelectorsIDs } from '../../../../Views/Wallet/WalletView.testIds';
+import { usePerpsNavigationHandlers } from './hooks/usePerpsNavigationHandlers';
+
+const PerpsSectionTrendingOnly = forwardRef<
+  SectionRefreshHandle,
+  PerpsSectionProps
+>(
+  (
+    {
+      sectionIndex,
+      totalSectionsLoaded,
+      sectionName: sectionNameOverride,
+      titleOverride,
+    },
+    ref,
+  ) => {
+    const sectionViewRef = useRef<View>(null);
+    const title = titleOverride ?? strings('homepage.sections.perpetuals');
+    const analyticsName = sectionNameOverride ?? HomeSectionNames.PERPS;
+    const trendingTransactionActiveAbTests =
+      useHomepageTrendingTransactionActiveAbTests();
+    const { handleViewAllPerps, handleViewMorePerps, handleTilePress } =
+      usePerpsNavigationHandlers({
+        trendingTransactionActiveAbTests,
+      });
+    const { marketsLoading, allCarouselMarkets, watchlistSymbolSet } =
+      usePerpsTrendingCarouselData({});
+    const carouselSymbols = useMemo(
+      () => allCarouselMarkets.map((m) => m.symbol),
+      [allCarouselMarkets],
+    );
+    const { sparklines, refresh: refreshSparklines } =
+      useHomepageSparklines(carouselSymbols);
+
+    useImperativeHandle(
+      ref,
+      () => ({
+        refresh: async () => {
+          refreshSparklines();
+        },
+      }),
+      [refreshSparklines],
+    );
+
+    const itemCount = allCarouselMarkets.length;
+    const { onLayout } = useHomeViewedEvent({
+      sectionRef: !marketsLoading && itemCount > 0 ? sectionViewRef : null,
+      isLoading: marketsLoading,
+      sectionName: analyticsName,
+      sectionIndex,
+      totalSectionsLoaded,
+      isEmpty: itemCount === 0,
+      itemCount,
+    });
+
+    if (!marketsLoading && itemCount === 0) {
+      return null;
+    }
+
+    return (
+      <View ref={sectionViewRef} onLayout={onLayout}>
+        <Box gap={3}>
+          <SectionHeader
+            title={title}
+            onPress={handleViewAllPerps}
+            testID={WalletViewSelectorsIDs.HOMEPAGE_SECTION_TITLE(
+              analyticsName,
+            )}
+          />
+          {marketsLoading ? (
+            <SectionRow>
+              <PerpsPositionSkeleton />
+            </SectionRow>
+          ) : (
+            <PerpsTrendingCarousel
+              markets={allCarouselMarkets}
+              watchlistSymbolSet={watchlistSymbolSet}
+              sparklines={sparklines}
+              onPressMarket={handleTilePress}
+              onPressViewMore={handleViewMorePerps}
+            />
+          )}
+        </Box>
+      </View>
+    );
+  },
+);
+
+PerpsSectionTrendingOnly.displayName = 'PerpsSectionTrendingOnly';
+
+export default PerpsSectionTrendingOnly;
