+++ app/components/Views/Homepage/Sections/Perpetuals/constants.ts
@@ -1,0 +1,3 @@
+export const MAX_ITEMS = 5;
+export const MAX_TRENDING_MARKETS = 5;
+export const HOMEPAGE_THROTTLE_MS = 5000;
